feat: added gamemode and platform filtering on all player data endpoint
