+++ app/players/parsers/player_career_parser.py
@@ -54,7 +54,13 @@ class PlayerCareerParser(BasePlayerParser):
         if kwargs.get("summary"):
             return self.data.get("summary")
 
-        return self._filter_stats(**kwargs) if kwargs.get("stats") else self.data
+        if kwargs.get("stats"):
+            return self._filter_stats(**kwargs)
+
+        return {
+            "summary": self.data["summary"],
+            "stats": self._filter_all_stats_data(**kwargs),
+        }
 
     def _filter_stats(self, **kwargs) -> dict:
         filtered_data = self.data["stats"] or {}
@@ -72,6 +78,7 @@ class PlayerCareerParser(BasePlayerParser):
                 platform = possible_platforms[0]
             else:
                 return {}
+
         filtered_data = filtered_data.get(platform) or {}
         if not filtered_data:
             return {}
@@ -88,6 +95,44 @@ class PlayerCareerParser(BasePlayerParser):
             for hero_key, statistics in filtered_data.items()
             if not hero_filter or hero_filter == hero_key
         }
+
+    def _filter_all_stats_data(self, **kwargs) -> dict:
+        stats_data = self.data["stats"] or {}
+        platform_filter = kwargs.get("platform")
+        gamemode_filter = kwargs.get("gamemode")
+
+        # Return early if no platform or gamemode is specified
+        if not platform_filter and not gamemode_filter:
+            return stats_data
+
+        # Extract platform-specific data and update stats data
+        # to contain only platform-specific information
+        if platform_filter:
+            stats_data = {
+                platform_key: (
+                    platform_data if platform_key == platform_filter else None
+                )
+                for platform_key, platform_data in stats_data.items()
+            }
+
+        # Extract gamemode-specific data within the remaining platforms
+        # and update stats data to contain only gamemode-specific information
+        if gamemode_filter:
+            stats_data = {
+                platform_key: (
+                    {
+                        gamemode_key: (
+                            gamemode_data if gamemode_key == gamemode_filter else None
+                        )
+                        for gamemode_key, gamemode_data in platform_data.items()
+                    }
+                    if platform_data is not None
+                    else None
+                )
+                for platform_key, platform_data in stats_data.items()
+            }
+
+        return stats_data
 
     def parse_data(self) -> dict:
         # We must check if we have the expected section for profile. If not,
