+++ app/players/parsers/search_data_parser.py
@@ -48,8 +48,8 @@ class SearchDataParser(JSONParser, ABC):
         return {self.data_type: data_value}
 
     def get_blizzard_url(self, **kwargs) -> str:
-        player_battletag = kwargs.get("player_id").replace("-", "#")
-        return f"{super().get_blizzard_url(**kwargs)}/{player_battletag}"
+        player_name = kwargs.get("player_id").split("-")[0]
+        return f"{super().get_blizzard_url(**kwargs)}/{player_name}/"
 
     def retrieve_data_value(self, player_data: dict) -> str | None:
         # If the player doesn't have any related data, directly return nothing here
